x86: Add MARCH_NATIVE configuration
Add a "Native" option under

Processor type and features
 => Processor family

If this option is set, the -march=native flag will be passed to the
compiler, telling it to target the CPU that we're currently running on.
It's a great option to have when building the kernel in the same machine
that will be using it.